+++ mn/src/mn/linux/Process.cpp
@@ -9,13 +9,13 @@ namespace mn
 	Process
 	process_id()
 	{
-		return Process{ getpid() };
+		return Process{ (uint64_t)getpid() };
 	}
 
 	Process
 	process_parent_id()
 	{
-		return Process{ getppid() };
+		return Process{ (uint64_t)getppid() };
 	}
 
 	bool
